requirements.txt
opencv-python
torch
numpy
Pillow
torchvision
huggingface_hub
Flask
Werkzeug